apps/backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
